requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Babel==4.0.0
Authlib==1.3.0
requests==2.31.0
psycopg[binary]==3.3.3
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.3
